% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{Analysis}
\alias{Analysis}
\title{A little analyisis that need to be deeper
Calculate the RollingMEan and Analyze keys statistics}
\usage{
Analysis(z)
}
\arguments{
\item{variable}{(usually, multi table xts)}
}
\description{
A little analyisis that need to be deeper
Calculate the RollingMEan and Analyze keys statistics
}
\examples{
Summary <- Analysis(ClosingPrices)
}
