% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modeldf.R
\name{Modeldf}
\alias{Modeldf}
\title{Transform an xts in a dataframe.
Useful with time series from GetSymbol
Applying getSymbols retrive quotation from Yahoo.Finance and choose the Close Price column.}
\usage{
Modeldf(z)
}
\arguments{
\item{xts}{objects from Yahoo Finance}
}
\value{
datframe
}
\description{
Transform an xts in a dataframe.
Useful with time series from GetSymbol
Applying getSymbols retrive quotation from Yahoo.Finance and choose the Close Price column.
}
\examples{
df <- Modeldf(ClosingPrices)
}
