% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot_Ticker}
\alias{Plot_Ticker}
\title{Plot your Tickers with ggplot2
Useful with time series from GetSymbol
Closing Prices and date up to date}
\usage{
Plot_Ticker(Dataset, Tckr1, Tckr2)
}
\arguments{
\item{dataset, }{Tickr1, Tickr2}
}
\value{
plot
}
\description{
Plot your Tickers with ggplot2
Useful with time series from GetSymbol
Closing Prices and date up to date
}
\examples{
Plot_Ticker(df, df$AAPL.Close, df$IBM.Close)
}
