% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESume.R
\name{Resume}
\alias{Resume}
\title{Save into Close Price of Stocks
#Merge and clean data.}
\usage{
Resume(x, y)
}
\arguments{
\item{NameTckr}{Yahoo Finance}
}
\value{
xts object, Stock close price
}
\description{
Save into Close Price of Stocks
#Merge and clean data.
}
\examples{
ClosingPrices <- Resume(Tickr1, Tickr2)
}
